refactor(go-template): extract spread-bag codegen into spread/spread-codegen.ts
Move the ten spread / object-map codegen methods out of the adapter
class into pure free functions (Roadmap B): two exported entry points
(collectSpreadSlots, buildSpreadInitializer) plus eight module-local
helpers. They read only state.restPropsName / state.usesFmt and
parseLiteralExpression — no new GoEmitContext member. The now-unused
parseRecordIndexAccess import is dropped from the adapter.

Output byte-identical, verified by the adapter unit (552) and
adapter-tests conformance (786) suites; tsgo + biome clean.
